Fix merging of adjacent URLs in ace-link--eww-collect
When there was no space between adjacent URLs, the old implementation of
ace-link--eww-collect merged them.  For example, this URL on the Common Lisp
HyperSpec was a problem:

http://www.lispworks.com/documentation/HyperSpec/Front/X_Master.htm

The images at the bottom of the page are links and were all merged by the old
implementation.

Make `ace-link--eww-collect' handle newlines more intelligently

Deal with links consisting solely of newlines

Fixes #64