Fix broken ace-link-compilation command
The problem here was that #64 caused a regression by changing the text property
from 'help-echo to 'shr-url.  This made sense for ace-link-eww but caused
problems for the other context that ace-link--eww-collect was used
in (ace-link-compilation).  I've fixed this by adding an optional argument that
can be used to set the text property.

It may be worth renaming this function since it is not just used for eww but
I've left it as is for now.