Revert "ace-link.el (ace-link-commit): Add"
This reverts commit 483d0ea9d1e13884f13e54093b41082884325878. The
major-mode named `magit-commit-mode` has been renamed in the magit to
`magit-revision-mode` in 2014 in
df088b28fa4296b0cc9d72f6b42cc42d232c0150:

  rename magit-commit-mode to magit-revision-mode

Also, the code in `ace-link-commit` uses functions that don't exist
anymore such as `magit-goto-next-section' and `magit-section-end'.

Instead of fixing all of that, I decided to get rid of the code that
no one is using anyway.